requirements.txt
# MONAI 및 PyTorch
torch
torchvision
monai[all]

# 데이터 처리 및 시각화
numpy
pandas
opencv-python
matplotlib
seaborn
scikit-learn
pydicom

# Grad-CAM (설명 가능 AI)
torchcam

# Streamlit 대시보드
streamlit
Pillow

# 기타 유틸리티
tqdm
